Replace goal action/reducer tests with integration test
